Added new API for adding streets with unit testing

+++ netherlands-parking-system/src/main/java/com/parking/nl/service/TariffService.java
@@ -1,9 +1,10 @@
 package com.parking.nl.service;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface TariffService {
 
-    public Map<String, Integer> loadTariffMetaData();
+    public Map<String, BigDecimal> loadTariffMetaData();
 
 }
