+++ netherlands-parking-system/src/main/java/com/parking/nl/validator/StreetValidator.java
@@ -7,6 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -21,7 +22,7 @@ public class StreetValidator{
     }
 
     public void validate(List<UnregisteredVehiclesRequest> unregisteredVehiclesRequest) {
-        Map<String, Integer> parkingTariffMetaData = tariffService.loadTariffMetaData();
+        Map<String, BigDecimal> parkingTariffMetaData = tariffService.loadTariffMetaData();
         List<String> streetNames = unregisteredVehiclesRequest.stream()
                 .map(UnregisteredVehiclesRequest::getStreetName).filter(streetName -> !parkingTariffMetaData.containsKey(streetName)).toList();
         if (!streetNames.isEmpty()) {
@@ -30,7 +31,7 @@ public class StreetValidator{
         }
     }
     public void validate(String streetName){
-        Map<String, Integer> parkingTariffMetaData = tariffService.loadTariffMetaData();
+        Map<String, BigDecimal> parkingTariffMetaData = tariffService.loadTariffMetaData();
         if(!parkingTariffMetaData.containsKey(streetName)){
             log.error("Street name  is not found in Street:",streetName);
             throw new InvalidInputException("Street name is not found");
